chore: Resources split into two files
