+++ src/components/Resource/index.js
@@ -1,26 +1,21 @@
 import React from 'react';
 import { Link } from 'gatsby';
-
 import { ArrowRight } from 'react-feather';
 
 import styles from './resource.module.css';
 
-const index = ({ node, currentPageId }) => {
-  const { id } = node;
-  const { title } = node.frontmatter;
+const index = ({ node }) => {
+  const { title, description } = node.frontmatter;
   const { slug } = node.fields;
 
   return (
-    <Link
-      to={slug}
-      className={`${styles.resource} ${
-        currentPageId === id ? styles.active : ''
-      }`}
-      aria-label={title}
-    >
+    <Link to={slug} className={styles.resource} aria-label={title}>
       <article key={slug}>
-        <h3>{title}</h3>
-        {currentPageId === id ? <ArrowRight size={16} /> : null}
+        <div>
+          <h3>{title}</h3>
+          <ArrowRight size={16} />
+        </div>
+        <p>{description}</p>
       </article>
     </Link>
   );
