+++ src/components/Resource/resource.module.css
@@ -1,22 +1,55 @@
-.resource {
+.resource,
+.resourceLink {
   text-decoration: none;
   color: var(--primary-foreground);
   position: relative;
   cursor: pointer;
 }
 
-.resource article {
+.resource {
+  display: block;
+  padding: 1rem;
+}
+
+.resource div {
+  display: flex;
+  align-items: center;
+}
+
+.resource h3,
+.resource svg {
+  margin-bottom: 0.5rem;
+  margin-right: 0.5rem;
+}
+
+.resource svg {
+  display: none;
+}
+
+.resource p {
+  margin-bottom: 0;
+}
+
+.resource:hover {
+  background: var(--foreground-min);
+}
+
+.resource:hover svg {
+  display: block;
+}
+
+.resourceLink article {
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 0.625rem 0.5rem;
 }
 
-.resource:hover {
+.resourceLink:hover {
   cursor: pointer;
 }
 
-.resource h3 {
+.resourceLink h3 {
   line-height: 1.3;
   font-size: 1.125rem;
   margin-bottom: 0rem;
@@ -27,19 +60,19 @@
   border-radius: var(--border-radius);
 }
 
-.resource:hover:not(.active) {
+.resourceLink:hover:not(.active) {
   background: var(--foreground-min);
 }
 
-.resource:hover:after:not(.active) {
+.resourceLink:hover:after:not(.active) {
   background: var(--foreground-hover);
 }
 
-.resource svg path {
+.resourceLink svg path {
   fill: var(--primary-foreground);
 }
 
-.resource:not(.active):not(:hover) svg {
+.resourceLink:not(.active):not(:hover) svg {
   display: none;
 }
